split out multfilterproxymodel
